+++ src/servers/reducers/selectedServer.js
@@ -1,9 +1,10 @@
 import ShlinkApiClient from '../../api/ShlinkApiClient';
 import ServersService from '../../servers/services/ServersService';
 import { resetShortUrlParams } from '../../short-urls/reducers/shortUrlsListParams'
+import { curry } from 'ramda';
 
-const SELECT_SERVER = 'shlink/selectedServer/SELECT_SERVER';
-const RESET_SELECTED_SERVER = 'shlink/selectedServer/RESET_SELECTED_SERVER';
+export const SELECT_SERVER = 'shlink/selectedServer/SELECT_SERVER';
+export const RESET_SELECTED_SERVER = 'shlink/selectedServer/RESET_SELECTED_SERVER';
 
 const defaultState = null;
 
@@ -20,7 +21,7 @@ export default function reducer(state = defaultState, action) {
 
 export const resetSelectedServer = () => ({ type: RESET_SELECTED_SERVER });
 
-export const selectServer = serverId => dispatch => {
+export const _selectServer = (ShlinkApiClient, ServersService, serverId) => dispatch => {
   dispatch(resetShortUrlParams());
 
   const selectedServer = ServersService.findServerById(serverId);
@@ -31,3 +32,4 @@ export const selectServer = serverId => dispatch => {
     selectedServer
   })
 };
+export const selectServer = curry(_selectServer)(ShlinkApiClient, ServersService);